refactor: viewResolver 역할 나누기
